+++ Repositories/EngineRepository.cs
@@ -16,7 +16,7 @@ namespace GameDatabase.Repositories
 
         public async Task<IEnumerable<Engine>> GetEnginesAsync()
         {
-            return await _context.Engines.OrderBy(e => e.EngineId).ToListAsync();
+            return await _context.Engines.Where(e => !e.IsDeleted).OrderBy(e => e.EngineId).ToListAsync();
         }
 
         public async Task<Engine> GetByIdAsync(int id)
@@ -38,17 +38,24 @@ namespace GameDatabase.Repositories
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(int id)
+        public async Task DeleteAsync(int id, string? deletedBy, bool isHardDelete)
         {
             var engine = await _context.Engines.FindAsync(id);
 
             if(engine != null)
             {
-                _context.Engines.Remove(engine);
-
+                if (isHardDelete)
+                {
+                    _context.Engines.Remove(engine);
+                }
+                else
+                {
+                    engine.IsDeleted = true;
+                    engine.DeletedBy = deletedBy;
+                    engine.DeletedAt = DateTime.UtcNow;
+                }
                 await _context.SaveChangesAsync();
             }
         }
-
     }
 }
